Refactor to use new UserEntityInterface
