+++ module/VuFind/src/VuFind/Db/Entity/AccessTokenInterface.php
@@ -38,7 +38,7 @@ namespace VuFind\Db\Entity;
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     https://vufind.org/wiki/development:plugins:database_gateways Wiki
  */
-interface AccessTokenInterface
+interface AccessTokenInterface extends EntityInterface
 {
     /**
      * Set user ID.
